requirements.txt
google-genai
efinance
pandas
plotly
jinja2
python-dotenv